+++ src/Components/App/App.js
@@ -1,6 +1,6 @@
 import './App.scss';
 import Header from '../Header/Header';
-import MovieCard from '../MovieCard/MovieCard';
+import MovieCard from '../MoviePoster/MoviePoster';
 import MoviesContainer from '../MoviesContainer/MoviesContainer';
 import movieData from './movieData';
 
@@ -8,7 +8,7 @@ function App() {
   return (
     <div className='App'>
       <Header />
-      <MoviesContainer movies={movieData.movies}/>
+      <MoviesContainer movies={movieData.movies} />
       Hello world
     </div>
   );
