+++ src/Components/Header/Header.scss
@@ -19,5 +19,5 @@ h1{
   font-family: 'showTime', sans-serif;
   font-size: 4rem;
   color: gold;
-
+  
 }
