+++ src/Components/Homepage/Homepage.js
@@ -1,0 +1,17 @@
+import './MoviesContainer.css';
+import MovieCard from '../MoviePoster/MoviePoster';
+
+function MoviesContainer({ movies }) {
+  const movieCards = movies.map((movie) => (
+    <MovieCard
+      key={movie.id}
+      id={movie.id}
+      posterPath={movie.poster_path}
+      title={movie.title}
+    />
+  ));
+
+  return <div className='moviesContainer'>{movieCards}</div>;
+}
+
+export default MoviesContainer;
