+++ src/Components/MovieDetail/MovieDetail.js
@@ -1,0 +1,5 @@
+// add a an onClick to the movie poster 
+// Add a new state for movie details 
+  // state starts out as 0
+    // conditional logic --> if movieDetails = false --> display homePage
+    // else if movie details = true --> don't display homePage
